This commit speeds up tests.
With this commit, we only load the virus engine once in our specs.

This also changes the way we determine if the virname variable _should_
contain a valid virus name. Accessing the virname when there is no virus
name segfaults, so we want to make sure that we avoid that at all costs.